script for collecting statistincs functional

+++ script/collect_statistics2.sh
@@ -4,7 +4,7 @@
 #check arguments
 case $# in
     3);;
-    *) echo "wrong number of argument! 1st: path to folder containing maps. 2nd: path to file to save statistics. 3rd: path to folder containg rosbags to be played."  1>&2
+    *) echo "wrong number of argument! 1st: number of submaps. 2nd: path to file to save statistics. 3rd: path to folder containg folders with maps to collect statistic from."  1>&2
         exit -1
         ;;
 esac
@@ -16,18 +16,8 @@ rosrun stroll_bearnav map_match_info_listener $2 &
 P2=$!
 
 
-roslaunch stroll_bearnav evaluate.launch folder:=$1 &            	 
-P1=$!
 
 
-
-
-#rosrun dynamic_reconfigure dynparam set /navigator matchingRatio 1.0
-#rosrun dynamic_reconfigure dynparam set /feature_extraction detector 2
-#rosrun dynamic_reconfigure dynparam set /feature_extraction descriptor 2
-
-#rostopic pub -1 /map_preprocessor/goal stroll_bearnav/loadMapActionGoal '{ header: { seq: 1, stamp: now , frame_id: ""}, goal_id: { stamp: now, id: "/Action_client_loader-1-0.000"}, goal: {prefix: "day_hostibejk_0"}}' &
-#rostopic echo /map_preprocessor/feedback -n 1
 if [ ! -e $2 ]; then
     echo "File not found! and creating new for statistics"
     touch $2
@@ -40,22 +30,22 @@ TXT_FILES=( `ls` )
 for i in ${TXT_FILES[*]}
 do
 	
-	end=${i##*.}
-	if [ "$end" = "bag" ]; then
-		rosservice call setDistance "distance: 0.0"
-        echo "playing rosbag $i"
-        rostopic pub -1 /navigator/goal stroll_bearnav/navigatorActionGoal '{ header: { seq: 1, stamp: now, frame_id: ""}, goal_id: { stamp: now, id: "/Action_client_navigator-1-0.000"}, goal: {traversals: 0}}' 
-        rosbag play $i --clock &
-	    P4=$!
+	if [[ -d $i ]]; then
+        echo "setting time for: $i" #prida podminku pro i je int
+		rosrun dynamic_reconfigure dynparam set /listener currentTime $i
+	
+        echo "using map in sub folder: $i"
+        
+		roslaunch stroll_bearnav evaluate.launch folder_view:=$3/$i/ &            	 
+		P1=$!							
 	    
-		rostopic echo navigationInfo/histogram -n 3
-		#rostopic pub -1 /navigator/cancel actionlib_msgs/GoalID '{ stamp: now, id: "/Action_client_navigator-1-0.000"}' 
-	    kill $P4
+		rostopic echo navigationInfo/histogram -n $1
+	    kill $P1
+		sleep 4s
     fi 
 	
 done
 
 kill -2 $P2
-kill $P1
 
 exit 0
